// apps/journeys-admin/__generated__/GetMyMuxVideo.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: GetMyMuxVideo
// ====================================================

export interface GetMyMuxVideo_getMyMuxVideo {
  __typename: "MuxVideo";
  id: string;
  showGeneratedSubtitles: boolean | null;
}

export interface GetMyMuxVideo {
  getMyMuxVideo: GetMyMuxVideo_getMyMuxVideo;
}

export interface GetMyMuxVideoVariables {
  id: string;
}